Add fuzz tests to `manticore`
Unfortunately, fuzz tests involving serializing types with lifetimes in
them don't work yet, because we can't implement `Arbitrary` for
`&'static [u8]`.

Adding support will require conditional compilation that replaces
`&'a [u8]` with `Cow<'a, [u8]>` when the `arbitrary-derive` featyre is
enable; we'll do this in a followup.

This CL also includes a small Python script for generating fuzz targets
for Cerberus commands (which are really boilerplate-ey).

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>