[util] Add `mem::Ref`, a fuzz-safe reference
`mem::Ref` is just a `&T` under the hood, unless the fuzzing feature
`arbitrary-derive` is enabled, in which case it becomes a `Cow<'_, T>`.
This allows us to implement `Arbitrary` on `'static` variants of
`manticore::protocol` types that have lifetime parameters, while
ensuring that other fuzzed code behaves identically to the baseline.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>