[flash] Replace FlashZero with read_direct()
It turns out that FlashZero was too restrictive, since it coupled arena
semantics too closely with a flash object. The new API is similar in
spirit, but allows for different arenas to be used with different
allocations, while also allowing for Flash implementations to optimize
whether they need to touch the arena at all.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>