[manifest] Add a test against a pre-baked manifest
This adds a pre-encoded manifest in src/manifest/testdata, which is used
to test conformance of the parser. This manifest is generated by the
same tooling the Microsoft implementation uses to generate its test
data.

Unfortunately, it seems that Manticore does not implement signature
verification in a consistent way, and resolving this may require
proposing some changes to the spec. For now, the part of the test that
checks the signature is disabled.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>