[manifest] Add a `Manager` type for managing manifest updates
This type is analogous to a similar class of manager types in the
Microsoft C implementation, whose primary purpose is to perform
the necessary checks for activating a manifest.

Right now, we expect the caller to track the "staging" area separately,
though it may wind up making sense to move that into `Manager` in the
fullness of time.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>