[server] Switch most PaRoT fields to dyn pointers
Most of these fields don't need to be type parameters: the traits are
guaranteed to be object safe, they are called infrequently, and most
consumers would use the dyn versions anyways for size.

This has the benefit of removing several type parameters that incurred a
high cost for a small benefit.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>